Separate out the key pair functionality
Move the key pair functionality into a separate file, to make it easier
for chip vendors to use their own key pair implementation.
